+++ clients/ctl/admin-ui/cypress/e2e/taxonomy.cy.ts
@@ -208,6 +208,13 @@ describe("Taxonomy management page", () => {
       );
       cy.getByTestId("input-recipients").should("contain", "marketing team");
       cy.getByTestId("input-recipients").should("contain", "dog shelter");
+      cy.getByTestId("input-legitimate_interest").within(() => {
+        cy.getByTestId("option-false").should("have.attr", "data-checked");
+        cy.getByTestId("option-true").click();
+        cy.getByTestId("option-true").should("have.attr", "data-checked");
+        cy.getByTestId("option-false").should("not.have.attr", "data-checked");
+      });
+
       cy.getByTestId("input-legitimate_interest_impact_assessment").should(
         "have.value",
         "https://example.org/legitimate_interest_assessment"
@@ -244,8 +251,7 @@ describe("Taxonomy management page", () => {
       cy.getByTestId("input-special_category").should("contain", "Select...");
       cy.getByTestId("input-recipients").should("contain", "Select...");
       cy.getByTestId("input-legitimate_interest_impact_assessment").should(
-        "have.value",
-        ""
+        "not.exist"
       );
     });
 
@@ -266,6 +272,12 @@ describe("Taxonomy management page", () => {
         cy.getByTestId("input-rights").should("contain", v);
       });
       cy.getByTestId("input-strategy").should("contain", "INCLUDE");
+      cy.getByTestId("input-automatic_decisions_or_profiling").within(() => {
+        cy.getByTestId("option-true").should("have.attr", "data-checked");
+        cy.getByTestId("option-false").click();
+        cy.getByTestId("option-false").should("have.attr", "data-checked");
+        cy.getByTestId("option-true").should("not.have.attr", "data-checked");
+      });
 
       // trigger a PUT
       cy.getByTestId("input-name").clear().type("foo");
@@ -273,7 +285,7 @@ describe("Taxonomy management page", () => {
       cy.wait("@putDataSubject").then((interception) => {
         const { body } = interception.request;
         const expected = {
-          automatic_decisions: true,
+          automatic_decisions_or_profiling: false,
           description:
             "An individual registered to voter with a state or authority.",
           fides_key: "citizen_voter",
@@ -291,7 +303,7 @@ describe("Taxonomy management page", () => {
       cy.getByTestId("item-Anonymous User").trigger("mouseover");
       cy.getByTestId("edit-btn").click();
       cy.getByTestId("input-rights").should("contain", "Select...");
-      cy.getByTestId("input-strategy").should("contain", "Select...");
+      cy.getByTestId("input-strategy").should("not.exist");
     });
 
     it("Can trigger an error", () => {