+++ clients/ctl/admin-ui/src/features/taxonomy/TaxonomyTabContent.tsx
@@ -59,7 +59,7 @@ const TaxonomyTabContent = ({ useTaxonomy }: Props) => {
     handleCreate: createEntity,
     handleEdit,
     handleDelete: deleteEntity,
-    extraFormFields,
+    renderExtraFormFields,
     transformEntityToInitialValues,
   } = useTaxonomy();
   const taxonomyNodes = useMemo(() => {
@@ -152,7 +152,7 @@ const TaxonomyTabContent = ({ useTaxonomy }: Props) => {
             labels={labels}
             onCancel={() => setEditEntity(null)}
             onSubmit={handleEdit}
-            extraFormFields={extraFormFields}
+            renderExtraFormFields={renderExtraFormFields}
             initialValues={transformEntityToInitialValues(editEntity)}
           />
         ) : null}
@@ -161,7 +161,7 @@ const TaxonomyTabContent = ({ useTaxonomy }: Props) => {
             labels={labels}
             onCancel={closeAddForm}
             onSubmit={createEntity}
-            extraFormFields={extraFormFields}
+            renderExtraFormFields={renderExtraFormFields}
             initialValues={transformEntityToInitialValues(
               DEFAULT_INITIAL_VALUES
             )}