+++ CHANGELOG.md
@@ -29,7 +29,7 @@ The types of changes are:
 
 ### Added
 
-* Added the ability to edit taxonomy fields via the UI [#977](https://github.com/ethyca/fides/pull/977)
+* Added the ability to edit taxonomy fields via the UI [#977](https://github.com/ethyca/fides/pull/977) [#1028](https://github.com/ethyca/fides/pull/1028)
 * New column `is_default` added to DataCategory, DataUse, DataSubject, and DataQualifier tables [#976](https://github.com/ethyca/fides/pull/976)
 * Added the ability to add taxonomy fields via the UI [#1019](https://github.com/ethyca/fides/pull/1019)
 * Added the ability to delete taxonomy fields via the UI [#1006](https://github.com/ethyca/fides/pull/1006)